+++ cli/down_test.go
@@ -21,6 +21,8 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/NethermindEth/sedge/configs"
+
 	"github.com/NethermindEth/sedge/internal/pkg/commands"
 	"github.com/NethermindEth/sedge/test"
 	log "github.com/sirupsen/logrus"
@@ -42,12 +44,12 @@ func buildDownTestCase(t *testing.T, caseName string, isErr bool) *downCmdTestCa
 		t.Fatalf("Can't build test case: %v", err)
 	}
 
-	dcPath := filepath.Join(configPath, "docker-compose-scripts")
+	dcPath := filepath.Join(configPath, configs.DefaultSedgeDataFolderName)
 	if err = os.Mkdir(dcPath, os.ModePerm); err != nil {
 		t.Fatalf("Can't build test case: %v", err)
 	}
 
-	err = test.PrepareTestCaseDir(filepath.Join("testdata", "down_tests", caseName, "docker-compose-scripts"), dcPath)
+	err = test.PrepareTestCaseDir(filepath.Join("testdata", "down_tests", caseName, configs.DefaultSedgeDataFolderName), dcPath)
 	if err != nil {
 		t.Fatalf("Can't build test case: %v", err)
 	}